Refactor: ScoreList를 ScoreStore로 리펙토링
- ScoreList class -> ScoreStore class
- Main 적용
